Make FitsFile.fptr private
The fitsfile and fitshdu files had to be merged for this to work.
Alternatively I could learn rust better...
